refactor(lex-babel): split markdown and html serializers into submodules
